requirements.txt
altgraph==0.17.4
Babel==2.14.0
packaging==23.2
pefile==2023.2.7
pillow==10.2.0
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
pywin32-ctypes==0.2.2
setuptools==69.0.3
tkcalendar==1.6.1
